Reuse avatar function by placing it in a "util"
Can be changed to a component later, just moving it here for now
